Try and generate a text signature template for both methods and functions.
